+++ Core/Core/View/Base/Webview/Models/WebviewInjection.swift
@@ -1,0 +1,46 @@
+//
+//  WebviewInjection.swift
+//  Core
+//
+//  Created by Vadim Kuznetsov on 4.01.24.
+//
+
+import WebKit
+public struct WebviewInjection: WebViewScriptInjectionProtocol {
+    public var id: String
+    public var script: String
+    public var messages: [WebviewMessage]?
+    public var injectionTime: WKUserScriptInjectionTime
+    init(
+        id: String,
+        script: String,
+        messages: [WebviewMessage]? = nil,
+        injectionTime: WKUserScriptInjectionTime = .atDocumentEnd
+    ) {
+        self.id = id
+        self.script = script
+        self.messages = messages
+        self.injectionTime = injectionTime
+    }
+    
+    public static func == (lhs: WebviewInjection, rhs: WebviewInjection) -> Bool {
+        lhs.id == rhs.id &&
+        lhs.script == rhs.script &&
+        lhs.injectionTime == rhs.injectionTime &&
+        lhs.messages == rhs.messages
+    }
+}
+
+public extension WebviewInjection {
+
+    static var surveyCSS: WebviewInjection {
+        SurveyCssInjection()
+            .webviewInjection()
+    }
+    
+    static var dragAndDropXss: WebviewInjection {
+        DragAndDropCssInjection()
+            .webviewInjection()
+    }
+
+}
