+++ Core/Core/View/Base/Webview/Models/WebviewMessage.swift
@@ -1,0 +1,16 @@
+//
+//  WebviewMessage.swift
+//  Core
+//
+//  Created by Vadim Kuznetsov on 4.01.24.
+//
+
+import WebKit
+public struct WebviewMessage: Equatable {
+    var name: String
+    var handler: (Any, WKWebView?) -> Void
+
+    public static func == (lhs: WebviewMessage, rhs: WebviewMessage) -> Bool {
+        lhs.name == rhs.name
+    }
+}
